+++ src/utils/api/fetchClaims.ts
@@ -6,12 +6,10 @@ import type {
 } from '@folio/stripes-acq-components';
 
 import {
-  fetchPieces,
-  fetchOrderLinesByIds,
-  fetchOrdersByIds,
   fetchOrganizationsByIds,
-  fetchReceivingTitlesByIds,
 } from '@folio/stripes-acq-components';
+
+import { fetchWrapperPieces } from './fetchWrapperPieces';
 
 interface DataShape {
   claims: ACQ.Claim[];
@@ -20,38 +18,25 @@ interface DataShape {
 
 export const fetchClaims = (httpClient: HTTPClient) => async (options: HTTPClientOptions): Promise<DataShape> => {
   const { signal } = options;
+  const { wrapperPieces, totalRecords } = await fetchWrapperPieces(httpClient)(options);
 
-  const { pieces, totalRecords } = await fetchPieces(httpClient)(options);
+  const organizationIdsSet = new Set<string>(wrapperPieces.map(({ vendorId }) => vendorId));
+  const { organizations } = await fetchOrganizationsByIds(httpClient)(
+    Array.from(organizationIdsSet), { signal },
+  );
 
-  const titleIdsSet = new Set(pieces.map(({ titleId }) => titleId));
-  const poLineIdsSet = new Set(pieces.map(({ poLineId }) => poLineId));
-
-  const [{ titles }, { poLines }] = await Promise.all([
-    fetchReceivingTitlesByIds(httpClient)(Array.from(titleIdsSet), { signal }),
-    fetchOrderLinesByIds(httpClient)(Array.from(poLineIdsSet), { signal }),
-  ]);
-
-  const orderIdsSet = new Set(poLines.map(({ purchaseOrderId }) => purchaseOrderId));
-  const { purchaseOrders } = await fetchOrdersByIds(httpClient)(Array.from(orderIdsSet), { signal });
-
-  const organizationIdsSet = new Set(purchaseOrders.map(({ vendor }) => vendor));
-  const { organizations } = await fetchOrganizationsByIds(httpClient)(Array.from(organizationIdsSet), { signal });
-
-  const purchaseOrdersDict = keyBy(purchaseOrders, 'id');
-  const poLinesDict = keyBy(poLines, 'id');
-  const titlesDict = keyBy(titles, 'id');
   const organizationsDict = keyBy(organizations, 'id');
 
   return {
-    claims: pieces.map((piece) => {
-      const vendor = organizationsDict[purchaseOrdersDict[poLinesDict[piece.poLineId].purchaseOrderId].vendor];
+    claims: wrapperPieces.map(({ piece, vendorId, poLine, title }) => {
+      const vendor = organizationsDict[vendorId];
 
       return {
         ...piece,
-        title: titlesDict[piece.titleId].title,
+        title: title.title,
         vendorCode: vendor.code,
         vendorName: vendor.name,
-        poLineNumber: poLinesDict[piece.poLineId].poLineNumber,
+        poLineNumber: poLine.poLineNumber,
       };
     }),
     totalRecords,