+++ src/views/Claiming/Claiming.tsx
@@ -55,9 +55,11 @@ import {
 import {
   ClaimingList,
   ClaimingListFilters,
+  GroupByOrgActionMenuItem,
   MarkUnreceivableModal,
 } from './components';
 import {
+  CLAIMING_HIDDEN_LIST_COLUMNS,
   CLAIMING_LIST_COLUMNS,
   CLAIMING_LIST_SORTABLE_FIELDS,
   FILTERS,
@@ -185,11 +187,25 @@ export const Claiming: React.FC = () => {
 
   const queryFilter = filters?.[SEARCH_PARAMETER];
   const pageTitle = queryFilter ? intl.formatMessage({ id: 'ui-claims.document.title.search' }, { query: queryFilter }) : null;
+  const sortByOrg = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+    if (sortingField !== CLAIMING_HIDDEN_LIST_COLUMNS.vendorId) {
+      changeSorting(e, { name: CLAIMING_HIDDEN_LIST_COLUMNS.vendorId });
+    }
+  };
 
   const renderActionMenu = ({ onToggle }: { onToggle: (e?: Event) => void }) => {
     return (
       <>
         <MenuSection>
+          <IfPermission perm="orders.wrapper-pieces.collection.get">
+            <GroupByOrgActionMenuItem
+              onClick={(e) => {
+                onToggle(e as unknown as Event | undefined);
+                sortByOrg(e);
+              }}
+            />
+          </IfPermission>
+
           <IfPermission perm="pieces.send-claims.collection.post">
             <SendClaimActionMenuItem
               disabled={!selectedRecordsLength}