+++ src/views/Claiming/components/index.tsx
@@ -1,3 +1,4 @@
 export { default as ClaimingList } from './ClaimingList';
 export { default as ClaimingListFilters } from './ClaimingListFilters';
+export * from './menu-items';
 export * from './modals';