+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 * [UICLAIM-4](https://folio-org.atlassian.net/browse/UICLAIM-4) Implement "Delay claim" action.
 * [UICLAIM-5](https://folio-org.atlassian.net/browse/UICLAIM-5) Implement "Unreceivable" action.
 * [UICLAIM-8](https://folio-org.atlassian.net/browse/UICLAIM-8) Adjust permissions for claiming.
+* [UICLAIM-2](https://folio-org.atlassian.net/browse/UICLAIM-2) Implement "Group by organization" action.